requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
gradio
idna==3.10
requests==2.32.4
streamlit
urllib3==2.4.0
python-dotenv
